[Cleanup] Allow Controller QueueElements to be excluded from a build
