+++ platformio.ini
@@ -335,7 +335,7 @@ board_build.flash_mode    = dout
 board_build.f_cpu         = ${common.board_build.f_cpu}
 board_upload.maximum_size = 786432
 build_unflags             = ${regular_platform.build_unflags}
-build_flags               = -Wl,-Tesp8266.flash.1m128.ld
+build_flags               = -Wl,-Tesp8266.flash.1m128.ld -DSIZE_1M
 
 [esp8285_1M]
 board                     = esp8285
