+++ src/_C001.ino
@@ -117,6 +117,8 @@ bool CPlugin_001(byte function, struct EventStruct *event, String& string)
   return success;
 }
 
+bool do_process_c001_delay_queue(int controller_number, const C001_queue_element& element, ControllerSettingsStruct& ControllerSettings);
+
 bool do_process_c001_delay_queue(int controller_number, const C001_queue_element& element, ControllerSettingsStruct& ControllerSettings) {
   WiFiClient client;
   if (!try_connect_host(controller_number, client, ControllerSettings))
