+++ src/_C003.ino
@@ -63,6 +63,8 @@ bool CPlugin_003(byte function, struct EventStruct *event, String& string)
   return success;
 }
 
+bool do_process_c003_delay_queue(int controller_number, const C003_queue_element& element, ControllerSettingsStruct& ControllerSettings);
+
 bool do_process_c003_delay_queue(int controller_number, const C003_queue_element& element, ControllerSettingsStruct& ControllerSettings) {
   bool success = false;
   char log[80];
