+++ src/_C004.ino
@@ -74,6 +74,8 @@ bool CPlugin_004(byte function, struct EventStruct *event, String& string)
   return success;
 }
 
+bool do_process_c004_delay_queue(int controller_number, const C004_queue_element& element, ControllerSettingsStruct& ControllerSettings);
+
 bool do_process_c004_delay_queue(int controller_number, const C004_queue_element& element, ControllerSettingsStruct& ControllerSettings) {
   WiFiClient client;
   if (!try_connect_host(controller_number, client, ControllerSettings))
