+++ src/_C009.ino
@@ -86,6 +86,8 @@ bool CPlugin_009(byte function, struct EventStruct *event, String& string)
 /*********************************************************************************************\
  * FHEM HTTP request
 \*********************************************************************************************/
+bool do_process_c009_delay_queue(int controller_number, const C009_queue_element& element, ControllerSettingsStruct& ControllerSettings);
+
 bool do_process_c009_delay_queue(int controller_number, const C009_queue_element& element, ControllerSettingsStruct& ControllerSettings) {
   WiFiClient client;
   if (!try_connect_host(controller_number, client, ControllerSettings))
