+++ src/_C018.ino
@@ -527,6 +527,8 @@ bool CPlugin_018(byte function, struct EventStruct *event, String& string)
   return success;
 }
 
+bool do_process_c018_delay_queue(int controller_number, const C018_queue_element& element, ControllerSettingsStruct& ControllerSettings);
+
 bool do_process_c018_delay_queue(int controller_number, const C018_queue_element& element, ControllerSettingsStruct& ControllerSettings) {
   bool  success = C018_data.txHexBytes(element.packed, ControllerSettings.Port);
   return success;
